backend/requirements.txt
flask
flask-cors
supabase
deepface
mtcnn
numpy
opencv-python-headless
scipy
python-dotenv
flask-bcrypt
